♻️ [Refac] API 관련 DTO 및 Router 리팩토링
