+++ ttakkeun/ttakkeun/Models/DTO/Diagnose/DiagnoseDeleteDTO.swift
@@ -1,0 +1,12 @@
+//
+//  DiagnoseDelete.swift
+//  ttakkeun
+//
+//  Created by Apple Coding machine on 8/15/25.
+//
+
+import Foundation
+
+struct DiagnoseDeleteResponse: Codable {
+    let deletionResult: String
+}
