+++ ttakkeun/ttakkeun/Models/DTO/Diagnose/DiagnosePointDTO.swift
@@ -1,0 +1,12 @@
+//
+//  DiagnosePointDTO.swift
+//  ttakkeun
+//
+//  Created by Apple Coding machine on 8/15/25.
+//
+
+import Foundation
+
+struct DiagnosePointResponse: Codable {
+    var point: Int
+}
