+++ ttakkeun/ttakkeun/Models/DTO/MyPage/MyPageUserInfoDTO.swift
@@ -1,0 +1,14 @@
+//
+//  MyPageUserInfoDTO.swift
+//  ttakkeun
+//
+//  Created by Apple Coding machine on 8/15/25.
+//
+
+import Foundation
+
+struct MyPageUserInfoResponse: Codable {
+    var memberId: Int
+    var username: String?
+    var email: String?
+}
