+++ ttakkeun/ttakkeun/Models/DTO/Record/RecordDeleteDTO.swift
@@ -1,0 +1,12 @@
+//
+//  RecordDeleteDTO.swift
+//  ttakkeun
+//
+//  Created by Apple Coding machine on 8/15/25.
+//
+
+import Foundation
+
+struct RecordDeleteResponse: Codable {
+    var message: String
+}
