--- ttakkeun/ttakkeun/Models/Request/Diagnosing/SelectedAnswerRequest.swift
@@ -1,24 +1,0 @@
-//
-//  SelectedAnswerData.swift
-//  ttakkeun
-//
-//  Created by 정의찬 on 11/18/24.
-//
-
-import Foundation
-
-struct SelectedAnswerRequest: Codable {
-    var petId: Int
-    var answers: [Int: [String]]
-    var etc: String?
-    
-    init(
-        petId: Int,
-        answers: [Int : [String]] = [:],
-        etc: String? = nil
-    ) {
-        self.petId = petId
-        self.answers = answers
-        self.etc = etc
-    }
-}
