--- ttakkeun/ttakkeun/Models/Response/Diagnosing/DeleteJournal.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteJournal.swift
-//  ttakkeun
-//
-//  Created by 정의찬 on 12/8/24.
-//
-
-import Foundation
-
-struct DeleteJournal: Codable {
-    var message: String
-}
