--- ttakkeun/ttakkeun/Models/Response/Diagnosing/DiagUserPoint.swift
@@ -1,12 +1,0 @@
-//
-//  DiagUserPoint.swift
-//  ttakkeun
-//
-//  Created by 정의찬 on 12/15/24.
-//
-
-import Foundation
-
-struct DiagUserPoint: Codable {
-    var point: Int
-}
