--- ttakkeun/ttakkeun/Models/Response/MyPage/MyInquiryResponse.swift
@@ -1,21 +1,0 @@
-//
-//  MyInquiryResponse.swift
-//  ttakkeun
-//
-//  Created by 황유빈 on 12/20/24.
-//
-
-import Foundation
-
-struct MyInquiryResponse: Equatable, Codable, Hashable, Identifiable {
-    var id = UUID()
-    var contents: String
-    var email: String
-    var inquiryType: String
-    var imageUrl: [String]
-    var created_at: String
-    
-    enum CodingKeys: String, CodingKey {
-        case contents, email, inquiryType, imageUrl, created_at
-    }
-}
