--- ttakkeun/ttakkeun/Models/Response/MyPage/UserInfoResponse.swift
@@ -1,14 +1,0 @@
-//
-//  MyPageResponse.swift
-//  ttakkeun
-//
-//  Created by 정의찬 on 12/18/24.
-//
-
-import Foundation
-
-struct UserInfoResponse: Codable {
-    var memberId: Int
-    var username: String?
-    var email: String?
-}
